User can delete list using trash bin by their list
I added an OnClickListener to all the trash bin image views and when it
is clicked the deleteList functino was called and it would pass the
position and the list object that is going to be deleted. The deleteList
function then uses a dialog to confirm the deletion of the list. And if
confirmed, the list is deleted from the DB