Added support to allow for ListItems to be updated
This bug fix allows for changes to made on the DB and reflected
correctly on the app. Also if the app causes a change to be written to
the DB the DB will receive the change and the changes will be handled
correctly in the arrayAdapter. Will need to be done to AllList as well